refactor: feedback domain에 vo 패키지 생성
