+++ backend/src/main/java/feedzupzup/backend/feedback/api/AdminFeedbackApi.java
@@ -1,7 +1,7 @@
 package feedzupzup.backend.feedback.api;
 
 import feedzupzup.backend.feedback.application.FeedbackOrderBy;
-import feedzupzup.backend.feedback.domain.ProcessStatus;
+import feedzupzup.backend.feedback.domain.vo.ProcessStatus;
 import feedzupzup.backend.feedback.dto.request.UpdateFeedbackCommentRequest;
 import feedzupzup.backend.feedback.dto.request.UpdateFeedbackSecretRequest;
 import feedzupzup.backend.feedback.dto.request.UpdateFeedbackStatusRequest;
