+++ backend/src/main/java/feedzupzup/backend/feedback/api/UserFeedbackApi.java
@@ -1,7 +1,7 @@
 package feedzupzup.backend.feedback.api;
 
 import feedzupzup.backend.feedback.application.FeedbackOrderBy;
-import feedzupzup.backend.feedback.domain.ProcessStatus;
+import feedzupzup.backend.feedback.domain.vo.ProcessStatus;
 import feedzupzup.backend.feedback.dto.request.CreateFeedbackRequest;
 import feedzupzup.backend.feedback.dto.response.CreateFeedbackResponse;
 import feedzupzup.backend.feedback.dto.response.LikeResponse;
