+++ backend/src/main/java/feedzupzup/backend/feedback/controller/AdminFeedbackController.java
@@ -3,7 +3,7 @@ package feedzupzup.backend.feedback.controller;
 import feedzupzup.backend.feedback.api.AdminFeedbackApi;
 import feedzupzup.backend.feedback.application.AdminFeedbackService;
 import feedzupzup.backend.feedback.application.FeedbackOrderBy;
-import feedzupzup.backend.feedback.domain.ProcessStatus;
+import feedzupzup.backend.feedback.domain.vo.ProcessStatus;
 import feedzupzup.backend.feedback.dto.request.UpdateFeedbackCommentRequest;
 import feedzupzup.backend.feedback.dto.response.UpdateFeedbackCommentResponse;
 import feedzupzup.backend.feedback.dto.response.UpdateFeedbackStatusResponse;
