+++ backend/src/main/java/feedzupzup/backend/feedback/domain/vo/ProcessStatus.java
@@ -1,4 +1,4 @@
-package feedzupzup.backend.feedback.domain;
+package feedzupzup.backend.feedback.domain.vo;
 
 public enum ProcessStatus {
     WAITING,
