+++ backend/src/main/java/feedzupzup/backend/feedback/domain/vo/UserName.java
@@ -1,4 +1,4 @@
-package feedzupzup.backend.feedback.domain;
+package feedzupzup.backend.feedback.domain.vo;
 
 import feedzupzup.backend.feedback.exception.FeedbackException;
 import feedzupzup.backend.global.response.ErrorCode;
