+++ backend/src/main/java/feedzupzup/backend/feedback/dto/request/CreateFeedbackRequest.java
@@ -1,11 +1,11 @@
 package feedzupzup.backend.feedback.dto.request;
 
 import feedzupzup.backend.category.domain.OrganizationCategory;
-import feedzupzup.backend.feedback.domain.Content;
+import feedzupzup.backend.feedback.domain.vo.Content;
 import feedzupzup.backend.feedback.domain.Feedback;
-import feedzupzup.backend.feedback.domain.PostedAt;
-import feedzupzup.backend.feedback.domain.ProcessStatus;
-import feedzupzup.backend.feedback.domain.UserName;
+import feedzupzup.backend.feedback.domain.vo.PostedAt;
+import feedzupzup.backend.feedback.domain.vo.ProcessStatus;
+import feedzupzup.backend.feedback.domain.vo.UserName;
 import io.swagger.v3.oas.annotations.media.Schema;
 import org.hibernate.validator.constraints.Length;
