+++ backend/src/main/java/feedzupzup/backend/feedback/dto/request/UpdateFeedbackStatusRequest.java
@@ -1,6 +1,6 @@
 package feedzupzup.backend.feedback.dto.request;
 
-import feedzupzup.backend.feedback.domain.ProcessStatus;
+import feedzupzup.backend.feedback.domain.vo.ProcessStatus;
 import io.swagger.v3.oas.annotations.media.Schema;
 
 @Schema(description = "피드백 상태 변경 요청")
