+++ backend/src/main/java/feedzupzup/backend/feedback/dto/response/AdminFeedbackListResponse.java
@@ -2,7 +2,7 @@ package feedzupzup.backend.feedback.dto.response;
 
 import feedzupzup.backend.feedback.domain.Feedback;
 import feedzupzup.backend.feedback.domain.FeedbackPage;
-import feedzupzup.backend.feedback.domain.ProcessStatus;
+import feedzupzup.backend.feedback.domain.vo.ProcessStatus;
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.time.LocalDateTime;
 import java.util.List;
@@ -67,7 +67,7 @@ public record AdminFeedbackListResponse(
                     feedback.getStatus(),
                     feedback.isSecret(),
                     feedback.getLikeCount(),
-                    feedback.getUserName(),
+                    feedback.getUserNameValue(),
                     feedback.getPostedAt().getPostedAt(),
                     feedback.getOrganizationCategory().getCategory().getKoreanName(),
                     feedback.getComment() != null ? feedback.getComment().getValue() : null
