+++ backend/src/main/java/feedzupzup/backend/feedback/dto/response/CreateFeedbackResponse.java
@@ -1,7 +1,7 @@
 package feedzupzup.backend.feedback.dto.response;
 
 import feedzupzup.backend.feedback.domain.Feedback;
-import feedzupzup.backend.feedback.domain.ProcessStatus;
+import feedzupzup.backend.feedback.domain.vo.ProcessStatus;
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.time.LocalDateTime;
 
@@ -35,7 +35,7 @@ public record CreateFeedbackResponse(
                 feedback.getContent().getValue(),
                 feedback.getStatus(),
                 feedback.isSecret(),
-                feedback.getUserName(),
+                feedback.getUserNameValue(),
                 feedback.getCreatedAt(),
                 feedback.getOrganizationCategory().getCategory().getKoreanName()
         );
