+++ backend/src/main/java/feedzupzup/backend/feedback/dto/response/UserFeedbackListResponse.java
@@ -1,7 +1,7 @@
 package feedzupzup.backend.feedback.dto.response;
 
 import feedzupzup.backend.feedback.domain.Feedback;
-import feedzupzup.backend.feedback.domain.ProcessStatus;
+import feedzupzup.backend.feedback.domain.vo.ProcessStatus;
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.time.LocalDateTime;
 import java.util.List;
@@ -70,7 +70,7 @@ public record UserFeedbackListResponse(
                     feedback.getStatus(),
                     feedback.isSecret(),
                     feedback.getLikeCount(),
-                    feedback.getUserName(),
+                    feedback.getUserNameValue(),
                     feedback.getPostedAt().getPostedAt(),
                     feedback.getOrganizationCategory().getCategory().getKoreanName(),
                     feedback.getComment() != null ? feedback.getComment().getValue() : null
