+++ backend/src/test/java/feedzupzup/backend/feedback/controller/UserFeedbackControllerE2ETest.java
@@ -1,8 +1,8 @@
 package feedzupzup.backend.feedback.controller;
 
 import static feedzupzup.backend.category.domain.Category.FACILITY;
-import static feedzupzup.backend.feedback.domain.ProcessStatus.CONFIRMED;
-import static feedzupzup.backend.feedback.domain.ProcessStatus.WAITING;
+import static feedzupzup.backend.feedback.domain.vo.ProcessStatus.CONFIRMED;
+import static feedzupzup.backend.feedback.domain.vo.ProcessStatus.WAITING;
 import static io.restassured.RestAssured.given;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.greaterThan;
