+++ backend/src/test/java/feedzupzup/backend/feedback/fixture/FeedbackFixture.java
@@ -1,12 +1,11 @@
 package feedzupzup.backend.feedback.fixture;
 
 import feedzupzup.backend.category.domain.OrganizationCategory;
-import feedzupzup.backend.feedback.domain.Comment;
-import feedzupzup.backend.feedback.domain.Content;
+import feedzupzup.backend.feedback.domain.vo.Content;
 import feedzupzup.backend.feedback.domain.Feedback;
-import feedzupzup.backend.feedback.domain.PostedAt;
-import feedzupzup.backend.feedback.domain.ProcessStatus;
-import feedzupzup.backend.feedback.domain.UserName;
+import feedzupzup.backend.feedback.domain.vo.PostedAt;
+import feedzupzup.backend.feedback.domain.vo.ProcessStatus;
+import feedzupzup.backend.feedback.domain.vo.UserName;
 
 public class FeedbackFixture {
